requirements.txt
flickrapi==2.4.0
Mastodon.py==1.8.1
mutagen==1.47.0
piexif==1.1.3
pillow==10.3.0
pybirdbuddy==0.0.17
python-dateutil==2.8.2
python-dotenv==1.0.1
python-graphql-client==0.4.3
python-magic==0.4.27
#python-magic-bin==0.4.14 nur für Windows
requests==2.31.0
